Attempt rearranging features to address bundle growth failure
