+++ src/common/entity/get_domain_features.ts
@@ -2,19 +2,19 @@ import { AITaskEntityFeature } from "../../data/ai_task";
 import { AlarmControlPanelEntityFeature } from "../../data/alarm_control_panel";
 import { AssistSatelliteEntityFeature } from "../../data/assist_satellite";
 import { CalendarEntityFeature } from "../../data/calendar";
-import { CameraEntityFeature } from "../../data/camera";
-import { ClimateEntityFeature } from "../../data/climate";
+import { CameraEntityFeature } from "../../data/feature/camera_entity_feature";
+import { ClimateEntityFeature } from "../../data/feature/climate_entity_feature";
 import { ConversationEntityFeature } from "../../data/conversation";
-import { CoverEntityFeature } from "../../data/cover";
+import { CoverEntityFeature } from "../../data/feature/cover_entity_feature";
 import { FanEntityFeature } from "../../data/fan";
 import { HumidifierEntityFeature } from "../../data/humidifier";
 import { LawnMowerEntityFeature } from "../../data/lawn_mower";
 import { LightEntityFeature } from "../../data/light";
 import { LockEntityFeature } from "../../data/lock";
-import { MediaPlayerEntityFeature } from "../../data/media-player";
+import { MediaPlayerEntityFeature } from "../../data/feature/media-player_entity_feature";
 import { NotifyEntityFeature } from "../../data/notify";
 import { RemoteEntityFeature } from "../../data/remote";
-import { SirenEntityFeature } from "../../data/siren";
+import { SirenEntityFeature } from "../../data/feature/siren_entity_feature";
 import { TodoListEntityFeature } from "../../data/todo";
 import { UpdateEntityFeature } from "../../data/update";
 import { VacuumEntityFeature } from "../../data/vacuum";
