+++ src/data/camera.ts
@@ -13,10 +13,7 @@ export const STREAM_TYPE_WEB_RTC = "web_rtc";
 
 export type StreamType = typeof STREAM_TYPE_HLS | typeof STREAM_TYPE_WEB_RTC;
 
-export enum CameraEntityFeature {
-  ON_OFF = 1,
-  STREAM = 2,
-}
+export { CameraEntityFeature } from "./feature/camera_entity_feature";
 
 interface CameraEntityAttributes extends HassEntityAttributeBase {
   model_name: string;
