+++ src/data/feature/camera_entity_feature.ts
@@ -1,0 +1,4 @@
+export enum CameraEntityFeature {
+  ON_OFF = 1,
+  STREAM = 2,
+}
