+++ src/data/feature/siren_entity_feature.ts
@@ -1,0 +1,7 @@
+export enum SirenEntityFeature {
+  TURN_ON = 1,
+  TURN_OFF = 2,
+  TONES = 4,
+  VOLUME_SET = 8,
+  DURATION = 16,
+}
