+++ src/data/media-player.ts
@@ -40,6 +40,9 @@ import type { MediaPlayerItemId } from "../components/media-player/ha-media-play
 import type { HomeAssistant, TranslationDict } from "../types";
 import { UNAVAILABLE } from "./entity/entity";
 import { isTTSMediaSource } from "./tts";
+import { MediaPlayerEntityFeature } from "./feature/media-player_entity_feature";
+
+export { MediaPlayerEntityFeature };
 
 interface MediaPlayerEntityAttributes extends HassEntityAttributeBase {
   media_content_id?: string;
@@ -80,29 +83,6 @@ export interface MediaPlayerEntity extends HassEntityBase {
     | "unknown"
     | "standby"
     | "buffering";
-}
-
-export enum MediaPlayerEntityFeature {
-  PAUSE = 1,
-  SEEK = 2,
-  VOLUME_SET = 4,
-  VOLUME_MUTE = 8,
-  PREVIOUS_TRACK = 16,
-  NEXT_TRACK = 32,
-
-  TURN_ON = 128,
-  TURN_OFF = 256,
-  PLAY_MEDIA = 512,
-  VOLUME_STEP = 1024,
-  SELECT_SOURCE = 2048,
-  STOP = 4096,
-  CLEAR_PLAYLIST = 8192,
-  PLAY = 16384,
-  SHUFFLE_SET = 32768,
-  SELECT_SOUND_MODE = 65536,
-  BROWSE_MEDIA = 131072,
-  REPEAT_SET = 262144,
-  GROUPING = 524288,
 }
 
 export type MediaPlayerBrowseAction = "pick" | "play";
